Convert BioCyc SmartTable

This function is used to perform specified conversion operations on BioCyc SmartTable.

Usage

biocyc_transform(session, table_id, transform_id, index = 0)

Arguments

session

httr::handle object, representing the established BioCyc session

table_id

string, unique identifier of SmartTable

transform_id

string, the ID of the transformation operation to be performed

index

integer, specifying the index of the column to be converted, the default is 0

Value

No return value, a success message will be printed when the function is executed successfully

Details

There are a total of 104 transformation operations, which can be obtained through cailab.utils:::biocyc_get_transformations(). Users can view the details of them in the transformations data by data("transformations").

See also

Examples

if (FALSE) { # \dontrun{
session <-biocyc_session("your_email@example.com", "your_password")
biocyc_transform(session, "your_table_id", "your_transform_id")
} # }