Install missing package and retry

I always encounter package not found error when I run render_book() for the first time. Normally, I check the error message and then rerun the command, and this make me annoy when I have to do it again and again. Therefore, I create this function to automate this operation.

Usage

install_missing_packages(expr = bookdown::render_book(), retry = 30)

Arguments

expr

an expression or command, i. e. bookdown::render_book()

retry

how many times should I retry. default is 30

Value

a message or a error that is not a 'packageNotFoundError'