The composition of community in a 8x8 plate (64 SynComs)

We use a fast, extensible method to generate multiple SynComs in every possible composition. The smallest unit is a 8x8 plate with 64 SynComs, which are all the combinations of six different strains. Based on this, one can generate as many as plates, and in that cases the combination of SynComs could be 64x2, 64x2x2, 64x2x2, and so on. For each time we add an extra strain, the number of total plates will double to fulfill all additional strain combinations. In other words, the possible combination of $n$ strain equals to $2^n$.

Usage

one_plate(
  plate_id = "P1",
  base_strain = "",
  add_strain = LETTERS[1:6],
  dim = c(8, 8),
  sep = "/",
  return_array = FALSE
)

Arguments

plate_id

a prefix in combination id

base_strain

the base strain or community in a single plate

add_strain

the id of added strains, which should contains 6 members

dim

c(8,8)

sep

separator

return_array

return an array if TRUE

Value

a tibble or an array

Examples

  one_plate(plate_id="",base_strain="",add_strain=LETTERS[1:6])
#> # A tibble: 64 × 2
#>    combination_id combination
#>    <chr>          <chr>      
#>  1 A1             A/B/C/D/E/F
#>  2 B1             A/B/C/D/E  
#>  3 C1             A/B/C/D/F  
#>  4 D1             A/B/C/D    
#>  5 E1             A/B/C/E/F  
#>  6 F1             A/B/C/E    
#>  7 G1             A/B/C/F    
#>  8 H1             A/B/C      
#>  9 A2             A/B/D/E/F  
#> 10 B2             A/B/D/E    
#> # ℹ 54 more rows