Run multiple sequence alignment with MUSCLE

Run multiple sequence alignment with MUSCLE

Usage

run_alignment_with_muscle(fasta_file, quiet = TRUE)

Arguments

fasta_file

sequence file(s) in fasta format, if a vector is provided, all files will be read.

quiet

silent MUSCLE progress

Value

a alignment object

Examples

file.path = system.file("extdata", "someORF.fa", package = "Biostrings")
run_alignment_with_muscle(file.path)
#> Runing multiple sequence alignment may take a while.
#> If you want to see the progress, set `quiet = FALSE`.
#> DNAMultipleAlignment with 7 rows and 8780 columns
#>      aln                                                    names               
#> [1] --------------------ACTTGT...GACCTTATTGTTGATAT-------- YAL001C TFC3 SGDI...
#> [2] --------------------------...AGAGTAAATTTTTTTCTATTCTCTT YAL002W VPS8 SGDI...
#> [3] CTTCATGTCAGCCTGCACTTCTGGGT...------------------------- YAL003W EFB1 SGDI...
#> [4] --------------------CACTCA...------------------------- YAL005C SSA1 SGDI...
#> [5] --------------------------...AAATATAATTTATGTGTGAACATAG YAL007C ERP2 SGDI...
#> [6] --------------------------...GAATGTACTTTT------------- YAL008W FUN14 SGD...
#> [7] --------------------------...------------------------- YAL009W SPO7 SGDI...